Add files via upload
Practice program to predict the class of new data. Using the Iris dataset we use the first 80% of the dataset as Training data and the last 20% as testing data. The testing data is a set of records with an unknown class  and the program will use the records to predict which class it belongs to.